+++ src/main/java/br/com/LeiloaBoardgames/repository/LeilaoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.LeiloaBoardgames.domain.Leilao;
 
 @Repository
-public interface LeilaoRepository extends JpaRepository<Leilao, Long> {
+public interface LeilaoRepository extends JpaRepository<Leilao, Integer> {
 
 }
